+++ api/src/main/java/com/example/api/http/screening.http
@@ -1,0 +1,3 @@
+### Screening API 테스트
+GET http://localhost:8080/api/screening
+Content-Type: application/json
