--- domain/movie-domain/src/main/java/com/example/moviedomain/entity/Movie.java
@@ -1,49 +1,0 @@
-package com.example.moviedomain.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "Movie")
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Movie {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Rating rating;
-
-    @Column(nullable = false)
-    private LocalDateTime releaseDate;
-
-    @Column(nullable = false)
-    private String thumbnail;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Genre genre;
-
-    @Column(nullable = false)
-    private Integer runtimeMinutes;
-
-    public enum Rating {
-        ALL, AGE_12, AGE_15, AGE_18
-    }
-
-    public enum Genre {
-        ACTION, ROMANCE, HORROR, SF
-    }
-
-
-}
