--- domain/movie-domain/src/main/java/com/example/moviedomain/entity/Screening.java
@@ -1,34 +1,0 @@
-package com.example.moviedomain.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "Screening")
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Screening {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private LocalDateTime startTime;
-
-    @Column(nullable = false)
-    private LocalDateTime endTime;
-
-    @ManyToOne
-    @JoinColumn(name = "movie_id", nullable = false)
-    private Movie movie;
-
-    @ManyToOne
-    @JoinColumn(name = "theater_id", nullable = false)
-    private Theater theater;
-}
