--- domain/movie-domain/src/main/java/com/example/moviedomain/repository/MovieRepository.java
@@ -1,7 +1,0 @@
-package com.example.moviedomain.repository;
-
-import com.example.moviedomain.entity.Movie;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MovieRepository extends JpaRepository<Movie,Long> {
-}
