--- domain/movie-domain/src/main/java/com/example/moviedomain/repository/TheaterRepository.java
@@ -1,7 +1,0 @@
-package com.example.moviedomain.repository;
-
-import com.example.moviedomain.entity.Theater;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TheaterRepository extends JpaRepository<Theater, Long> {
-}
