+++ domain/src/main/java/com/example/domain/screening/entity/Screening.java
@@ -1,0 +1,35 @@
+package com.example.domain.screening.entity;
+
+import com.example.domain.theater.entity.Theater;
+import jakarta.persistence.*;
+import lombok.*;
+import com.example.domain.movie.entity.Movie;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "Screening")
+@Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Screening {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private LocalDateTime startTime;
+
+    @Column(nullable = false)
+    private LocalDateTime endTime;
+
+    @ManyToOne
+    @JoinColumn(name = "movie_id", nullable = false)
+    private Movie movie;
+
+    @ManyToOne
+    @JoinColumn(name = "theater_id", nullable = false)
+    private Theater theater;
+}
