+++ infra/src/main/java/com/example/infra/movie/repository/MovieJpaRepository.java
@@ -1,0 +1,8 @@
+package com.example.infra.movie.repository;
+
+import com.example.domain.movie.entity.Movie;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MovieJpaRepository extends JpaRepository<Movie,Long> {
+
+}
