+++ infra/src/main/java/com/example/infra/seat/repository/SeatJpaRepository.java
@@ -1,0 +1,7 @@
+package com.example.infra.seat.repository;
+
+import com.example.domain.seat.entity.Seat;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SeatJpaRepository extends JpaRepository<Seat,Long> {
+}
