+++ infra/src/main/java/com/example/infra/theater/repository/TheaterJpaRepository.java
@@ -1,0 +1,8 @@
+package com.example.infra.theater.repository;
+
+import com.example.domain.theater.entity.Theater;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TheaterJpaRepository extends JpaRepository<Theater,Long> {
+
+}
